Update README.md
Cleaning up README docs across 18F repos, to make them uniform. I realize this is a deprecated repo, but added a link to gsa-advantage-scraper in case anyone comes across this. 

I tried to find a link to the nodemailer hack you mentioned in the intro paragraph and would recommend linking to that as well, if you feel like it would help someone who discovers this.

I would also recommend linking to the current, non-deprecated Communicart repo.

Mel